Check that a data value is consistent with its prior

Check that a data value is consistent with its prior

Usage

check_consistent(point, prior, p_error = 5e-04, p_warning = 0.025)

Arguments

point

(numeric) Data value to check

prior

list of distn, parama, paramb

p_error

(numeric) Probability value outside of which we raise an error

p_warning

(numeric) Probability value outside of which we raise a warning

Value

(c(no_error = , no_warning = ))